requirements.txt
SQLAlchemy==2.0.28
fastapi==0.110.0
alembic==1.13.1
python-dotenv==1.0.1
uvicorn==0.28.1
pydantic==2.6.4
pydantic-settings==2.2.1
passlib==1.7.4
PyJWT==2.8.0
asyncpg==0.29.0